DBMS/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
email-validator==2.1.0.post1
gunicorn==21.2.0
sentry-sdk[flask]==1.40.0
Flask-Caching==2.1.0
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.4.1
